Nov 10th/Tasks/requirements.txt
langgraph==0.0.35
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
langchain-openai==0.1.3
requests==2.31.0
streamlit==1.33.0
